BE FE: e-template/ create email_tp

+++ flatform-frontend/app/admin/users/actions.ts
@@ -1,8 +1,9 @@
 // app/admin/users/actions.ts
-import api from '@/lib/api'
+import api from "@/lib/api";
 
-export const getUsers = (params: any) => api.get('/users', { params })
-export const getUserById = (id: string) => api.get(`/users/${id}`)
-export const createUser = (data: any) => api.post('/users', data)
-export const updateUser = (id: string, data: any) => api.put(`/users/${id}`, data)
-export const disableUser = (id: string) => api.delete(`/users/${id}`)
+export const getUsers = (params: any) => api.get("/admin/users", { params });
+export const getUserById = (id: string) => api.get(`/admin/users/${id}`);
+export const createUser = (data: any) => api.post("/admin/users", data);
+export const updateUser = (id: string, data: any) =>
+  api.put(`/admin/users/${id}`, data);
+export const disableUser = (id: string) => api.delete(`/admin/users/${id}`);
